+++ src/app/js/admin/orgs.js
@@ -28,7 +28,7 @@
       getOrgs();
     }
 
-    $scope.$on('evtAuthenticateStateChanged', function(evt, masterAuth, user) {
+    $scope.$on('evtAuthenticateStateChanged', function() {
       if ($rootScope.userLoggedInIsAdmin()) {
         getOrgs();
       }