+++ src/app/js/admin/users.js
@@ -28,7 +28,7 @@
       getUsers();
     }
 
-    $scope.$on('evtAuthenticateStateChanged', function(evt, masterAuth, user) {
+    $scope.$on('evtAuthenticateStateChanged', function() {
       if ($rootScope.userLoggedInIsAdmin()) {
         getUsers();
       }